refactor: extract release semver planning
